pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pychromvar"
version = "0.0.2"
description = "A python package for chromVAR"
authors = [
    {name = "Zhijian Li", email = "lzj1769@gmail.com"}
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research"
]
requires-python = ">= 3.6"
dependencies = [
    "anndata",
    "mudata",
    "scanpy",
    "muon",
    "biopython",
    "MOODS-python",
    "pysam",
    "wget"
]

[project.urls]
Source = "https://github.com/lzj1769/pychromVAR"